requirements.txt
Flask==2.3.3
pymodbus==3.5.2
python-dotenv==1.0.0
flask-cors==4.0.0
